new files for Ev and Cat. Event plot function
