+++ obspy/core/event/__init__.py
@@ -23,7 +23,9 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
 from .base import *
-from .radpattern import *
+from .event import Event
+from .catalog import Catalog
+from .radpattern import plot_3drpattern, farfield_p, farfield_s
 
 if __name__ == '__main__':
     import doctest
